refactored auth service, fixed loadUserData

+++ Demo2/Core/User/List/ViewModel/UsersViewModel.swift
@@ -35,6 +35,8 @@ final class UsersViewModel: ObservableObject {
     
     @MainActor
     func removeUser(uid: String) async throws {
+        guard UserService.isCurrentUser(uid: uid) else { return }
+        
         self.users.removeValue(forKey: uid)
         try await UserService.removeUser(uid: uid)
         try Interface.auth().removeAccount()
